[tmp] Fall back to router reload for idled services
DCM cannot disable the health check dynamically. The API command exists but not used.
This causes a problem with idled services whose server slot is set with
the single endpoint - service's clusterIP.
However a health check sent to the service's clusterIP unidles the service.
In non DCM case this is mitigated by removing "check" keyword during
the config template rendering. Although in DCM case only the server slot is
updated with the clusterIP keeping the health check enabled.